Add newDocument.php. center forgetpw.php
I added a page for uploading new Documents. Both paper and electronic Docs are processed. Little javascript is used to load fields for either paper doc or for electronic doc. modified link on personal page to redirect to newDocument. Furthermore, the forgetpw block was in the top-left corner, added a div to move it to the center.